LayoutTests/crypto/webkitSubtle/aes-kw-key-manipulation.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test generating, importing and exporting keys for AES-KW. Test that they can't be used with another algorithm.");

jsTestIsAsync = true;

var extractable = true;

debug("Generating a key...");
crypto.webkitSubtle.generateKey({name: "aes-kw", length: 256}, extractable, ["encrypt", "decrypt", "wrapKey", "unwrapKey"]).then(function(result) {
    key = result;
    shouldBe("key.toString()", "'[object CryptoKey]'");
    shouldBe("key.type", "'secret'");
    shouldBe("key.algorithm.name", "'AES-KW'");
    shouldBe("key.algorithm.length", "256");

    debug("\nTesting that the key can't be used with AES-CBC...");
    iv = hexStringToUint8Array("000102030405060708090a0b0c0d0e0f");
    shouldThrow('crypto.webkitSubtle.encrypt({name: "aes-cbc", iv: iv}, key, hexStringToUint8Array("00"))');

    debug("\nExporting the key to raw...");
    return crypto.webkitSubtle.exportKey('raw', key);
}).then(function(result) {
    exportedKey = result;
    shouldBe("exportedKey.toString()", "'[object ArrayBuffer]'");
    debug("Importing it back...");
    return crypto.webkitSubtle.importKey('raw', exportedKey, "aes-kw", extractable, ["encrypt", "decrypt", "wrapKey", "unwrapKey"]);
}).then(function(result) {
    importedKey = result;

    shouldBe("importedKey.toString()", "'[object CryptoKey]'");
    shouldBe("importedKey.type", "'secret'");
    shouldBe("importedKey.algorithm.name", "'AES-KW'");
    shouldBe("importedKey.algorithm.length", "256");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
